+++ smash/__init__.py
@@ -12,7 +12,7 @@ from smash.io.setup_io import save_setup, read_setup
 from smash.io.mesh_io import save_mesh, read_mesh
 from smash.io.model_io import save_model, read_model
 from smash.io.model_ddt_io import save_model_ddt, read_model_ddt
-from smash.io.multi_model_io import save_smash_model_to_hdf5, load_hdf5_file, generate_smash_object_structure
+from smash.io.hdf5_io import save_smash_model_to_hdf5, load_hdf5_file, generate_smash_object_structure
 
 from smash.dataset.load import load_dataset
